requirements.txt
asgiref==3.7.2
attrs==23.2.0
cfgv==3.4.0
coverage==7.4.1
distlib==0.3.8
Django==5.0.1
django-filter==23.5
django-nested-admin==4.0.2
djangorestframework==3.14.0
drf-spectacular==0.27.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
filelock==3.13.1
flake8==7.0.0
gunicorn==21.2.0
identify==2.5.33
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
nodeenv==1.8.0
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.6.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==7.4.4
python-dotenv==1.0.0
python-monkey-business==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
virtualenv==20.25.0
